integration/requirements.txt
requests==2.18.4
retrying==1.3.3
nose2==0.6.5
python-openstackclient==3.2.0
os-client-config==1.22.0
waiting==1.3.0
ipython==5.1.0
awscli==1.10.65
docker-compose==1.13.0
flake8==3.0.4
pylint==1.8.1
boto3==1.4.4
pika==0.10.0
ipdb==0.10.1



# dependencies used to resolve unknown attribute of an object
# should be removed when removing openstack-client
oslo.config==3.21.0
oslo.i18n==3.11.0
oslo.utils==3.21.0
keystoneauth1==2.16.0
python-cinderclient==1.9.0
python-keystoneclient==3.8.0
cliff==2.3.0
osc-lib==1.2.0
stevedore==1.19.1
